requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
drf-social-oauth2==3.1.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.27.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
